Added `Loading` while checking for token.
There is an outstanding bug to address where a perpetual loading state can be achieved if no token is ever set -- even after the call is complete.

Solution: Add request `start` and `end` requests to `apiMiddleware` that will dispatch start/end types to the reducer and update a `STATE_KEY[RESOURCE]_IS_LOADING` constant in the store.