Updated `authOverride` to dynamically occur only on demo site.
This is so you can see the app in production if your back-end is not yet set up for authentication flows.

Also added a note to remove it when the app is capable of a full authentication flows, with backend calls.
Adjusted additional locations to remove in notes.


Tweaked line about mirage.